feat: Users data is stored in PostgreSQL-Prisma

+++ src/resources/users/users.controller.ts
@@ -10,10 +10,14 @@ import {
   UseInterceptors,
   ClassSerializerInterceptor,
   ParseUUIDPipe,
+  NotFoundException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
+
+import { User as UserModel } from '@prisma/client';
 
 @Controller('user')
 export class UsersController {
@@ -21,34 +25,54 @@ export class UsersController {
 
   @Post()
   @UseInterceptors(ClassSerializerInterceptor)
-  create(@Body() userDto: CreateUserDto) {
+  async create(@Body() userDto: CreateUserDto) {
     return this.usersService.create(userDto);
   }
 
   @Get()
   @UseInterceptors(ClassSerializerInterceptor)
-  findAll() {
+  async findAll() {
     return this.usersService.findAll();
   }
 
   @Get(':id')
   @UseInterceptors(ClassSerializerInterceptor)
-  findOne(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    return this.usersService.findOne(id);
+  async findOne(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
+    const user = await this.usersService.findOne(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    return user;
   }
 
   @Put(':id')
   @UseInterceptors(ClassSerializerInterceptor)
-  update(
+  async update(
     @Param('id', new ParseUUIDPipe({ version: '4' })) id: string,
     @Body() userDto: UpdateUserDto,
   ) {
-    return this.usersService.update(id, userDto);
+    const user = await this.usersService.findOne(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    } else {
+      const update = await this.usersService.update(id, userDto);
+      console.log(update);
+      if (!update) {
+        throw new ForbiddenException('Wrong password');
+      } else {
+        return update;
+      }
+    }
   }
 
   @Delete(':id')
   @HttpCode(204)
-  remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
-    return this.usersService.remove(id);
+  async remove(@Param('id', new ParseUUIDPipe({ version: '4' })) id: string) {
+    const user = await this.usersService.findOne(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    } else {
+      this.usersService.remove(id);
+    }
   }
 }
